added functionality to delete button
when user enters the id they want to delete, loop compares that to every line in data.txt. 
if the lines that were compared are not the same, add that line that it just read to a new file called temp.txt. also set a boolean called noIDFound to true as the id that is needed to be deleted has not been found yet.
else set noIDFound to false as we have now found the id that we want to delete.
new if statement that will be run if noIDFound is false that says it was a success
else say that the id was not found.
new try catch blocks will delete the existing data.txt as it is no longer the most updated version. then we will recreate it so that we can write all the lines of the temp.txt into the data.txt.