Fixed Issues with databaseBuild
The program now allows you to click the view button on program load if there is text in the data.txt file.
Previously, the program would force you to enter in at least one entry before you could view the full databaseBuild table.

The program also now allows you to enter in package names with spaces in them.
Previously the program split each component of the line on the text file by " | ". The space before and after the "|" would disallow the use of spaces in between program names.  The separator is now "<>".