unique ids and a few other small tweaks
added unique id functionality to the program. required a new boolean so that it only checks for duplicate ids if the user hasnt entered valid info in the first place.
basically the program just reads through data.txt to look for any matching ids from the one the user entered (these comparisons will work regardless of if the user has 0's in front of the id or not) and then if it matches it stops reading data.txt and tells the user they have entered a duplicate id and sets the isInformationCorrect boolean to false so that it doesnt add that duplicate id to data.txt.

deleted some System.out.println lines that are not required. 

there might have been some other little lines i deleted that had no use etc. however the program is now stable