sorting added
added sorting functionality. you are now able to sort by packageid, databaseid, alphabetical packagename, date arrived and weight (low to high)

added all of the code in databaseBuild to a frame so that it can be disposed of or opened up at any given moment. This was not possible beforehand as the frame that it was in was technically never declared/initialized.